requirements.txt
transformers>=4.43.0
torch>=2.1.0
torchvision
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.25.0